Consolidate all core engine logic under single dev_tools package
