+++ boomtick-pkg/cli/dev_tools/generate_aggregate_prs_workflow.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 import sys
 import os
-from tdw_services.orchestrator import Orchestrator
+from dev_tools.orchestrator import Orchestrator
 
 def main():
     orch = Orchestrator()
