+++ boomtick-pkg/cli/dev_tools/repair.py
@@ -9,7 +9,7 @@ import sys
 import json
 import re
 
-from tdw_services.orchestrator import Orchestrator
+from dev_tools.orchestrator import Orchestrator
 from dev_tools.utils import extract_failing_info
 
 MAX_RETRIES = 3
