--- boomtick-pkg/cli/dev_tools/td_cli.py
@@ -1,100 +1,0 @@
-#!/usr/bin/env python3
-"""
-td_cli.py - Project Developer CLI Shim
-
-This script is a thin wrapper around the unified tdw_services CLI.
-It maintains backward compatibility for existing scripts and CI workflows.
-"""
-
-import sys
-import os
-
-
-try:
-    from tdw_services.cli import cli
-    # Expose utilities for legacy tests
-    from dev_tools import utils
-    from dev_tools.utils import get_github_token, get_repo_name, get_gha_variable, CLIError, get_github_client
-
-    # Force tdw_services.orchestrator to use the same utility functions as td_cli
-    # so that legacy tests patching td_cli.get_github_client etc. will work.
-    import tdw_services.orchestrator
-    tdw_services.orchestrator.get_github_client = get_github_client
-    tdw_services.orchestrator.get_repo_name = get_repo_name
-    tdw_services.orchestrator.get_github_token = get_github_token
-    tdw_services.orchestrator.get_gha_variable = get_gha_variable
-
-    from tdw_services.orchestrator import Orchestrator
-    _orch = Orchestrator()
-    resolve_baseline = _orch.resolve_baseline
-
-    def handle_fix_ci(args):
-        # Support legacy test expectation for GITHUB_TOKEN
-        if not get_github_token():
-             raise CLIError("Missing GITHUB_TOKEN", code=401)
-
-        # Support test expectation for JULES_API_KEY
-        if not getattr(args, 'api_key', None) and not os.environ.get("JULES_API_KEY"):
-            raise CLIError("Missing JULES_API_KEY", code=401)
-
-        # Support legacy test expectation for repo name
-        if not get_repo_name():
-             raise CLIError("Could not determine repository name", code=400)
-
-        return _orch.fix_ci(
-            pr_number=getattr(args, 'pr_number', None),
-            branch=getattr(args, 'branch', None),
-            api_key=getattr(args, 'api_key', None),
-            dry_run=getattr(args, 'dry_run', True)
-        )
-
-    def handle_validate_issue(args):
-        return _orch.validate_issue(
-            issue_number=getattr(args, 'issue_number', None),
-            all_open=getattr(args, 'all_open', False),
-            post_comments=getattr(args, 'post_comments', False),
-            dry_run=getattr(args, 'dry_run', True)
-        )
-
-    def resolve_baseline(file_path, env_var, fallback):
-        # Match legacy test which mocks td_cli.get_gha_variable
-        val = get_gha_variable(env_var)
-        if val: return int(val)
-        return _orch.resolve_baseline(file_path, env_var, fallback)
-
-    def handle_audit_pr(args):
-        # Handle the case where args.pr_number might be a string like "null" from tests
-        pr_num = getattr(args, 'pr_number', None)
-        if pr_num in ["null", "", None]:
-             raise CLIError("Invalid PR number")
-        try:
-            pr_num = int(pr_num)
-        except (ValueError, TypeError):
-             raise CLIError("Invalid PR number format")
-
-        return _orch.audit_pr(
-            pr_num,
-            fetch=getattr(args, 'fetch', False),
-            audit=getattr(args, 'audit', False),
-            submit=getattr(args, 'submit', False),
-            cleanup=getattr(args, 'cleanup', False),
-            dry_run=getattr(args, 'dry_run', True),
-            event=getattr(args, 'event', None)
-        )
-except ImportError as e:
-    print(f"""Error: Could not import tdw_services or its dependencies.
-Details: {e}
-
-Troubleshooting:
-1. Ensure dependencies are installed: pip install -e boomtick-pkg/cli
-2. Ensure PYTHONPATH includes the dev-tools directory.
-   Example: export PYTHONPATH=$PYTHONPATH:$(pwd)/boomtick-pkg/cli""", file=sys.stderr)
-    if "pytest" not in sys.modules:
-        sys.exit(1)
-
-def main():
-    # click entry point automatically handles sys.argv
-    cli(obj={})
-
-if __name__ == "__main__":
-    main()
