+++ boomtick-pkg/cli/tests/test_parse_comment.py
@@ -1,5 +1,5 @@
 import pytest
-from tdw_services.orchestrator import Orchestrator
+from dev_tools.orchestrator import Orchestrator
 
 @pytest.fixture
 def orchestrator():
