+++ scripts/send-jules-impact.py
@@ -10,7 +10,7 @@ logger = logging.getLogger(__name__)
 # Add dev-tools to the python path since scripts might run from root
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'boomtick-pkg/cli')))
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'boomtick-pkg/cli/dev_tools')))
-from tdw_services.services.jules import JulesClient
+from dev_tools.services.jules import JulesClient
 
 def is_skipped_review(content: str) -> bool:
     lines = [line.strip() for line in content.splitlines() if line.strip()]
