+++ tests/dev-tools/test_ai_service.py
@@ -5,7 +5,7 @@ import os
 
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '../../dev-tools')))
 
-from tdw_services.services.ai_service import AIClient
+from dev_tools.services.ai_service import AIClient
 
 class TestAIService(unittest.TestCase):
     def setUp(self):
