+++ tests/dev-tools/test_command_handler.py
@@ -6,7 +6,7 @@ import os
 # Add dev-tools to path
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '../../dev-tools')))
 
-from tdw_services.handlers.command_handler import CommandHandler
+from dev_tools.handlers.command_handler import CommandHandler
 
 class TestCommandHandler(unittest.TestCase):
     def setUp(self):
