+++ tests/dev-tools/test_modern_cli.py
@@ -8,21 +8,21 @@ import json
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../../boomtick-pkg/cli/dev_tools')))
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../../boomtick-pkg/cli')))
 
-from tdw_services.cli import cli
+from dev_tools.cli import cli
 from click.testing import CliRunner
 
 class TestModernCLI(unittest.TestCase):
     def setUp(self):
         self.runner = CliRunner()
 
-    @patch('tdw_services.orchestrator.Orchestrator.validate_issue')
+    @patch('dev_tools.orchestrator.Orchestrator.validate_issue')
     def test_validate_issue_calls_orchestrator(self, mock_validate):
         mock_validate.return_value = {"status": "success", "issues": [], "total_findings": 0}
         result = self.runner.invoke(cli, ['gh', 'validate-issue', '--issue-number', '123'])
         self.assertEqual(result.exit_code, 0)
         mock_validate.assert_called_once()
 
-    @patch('tdw_services.orchestrator.Orchestrator.audit_pr')
+    @patch('dev_tools.orchestrator.Orchestrator.audit_pr')
     def test_audit_pr_calls_orchestrator(self, mock_audit):
         mock_audit.return_value = {"pr": 123}
         result = self.runner.invoke(cli, ['gh', 'audit-pr', '123', '--fetch'])
@@ -34,14 +34,14 @@ class TestModernCLI(unittest.TestCase):
         self.assertTrue(kwargs['fetch'])
         self.assertFalse(kwargs['audit'])
 
-    @patch('tdw_services.orchestrator.Orchestrator.analyze_file')
+    @patch('dev_tools.orchestrator.Orchestrator.analyze_file')
     def test_analyze_calls_orchestrator(self, mock_analyze):
         mock_analyze.return_value = "solid code"
         result = self.runner.invoke(cli, ['ai', 'analyze', 'README.md'])
         self.assertEqual(result.exit_code, 0)
         mock_analyze.assert_called_once_with('README.md')
 
-    @patch('tdw_services.orchestrator.Orchestrator.resolve_pr_conflicts')
+    @patch('dev_tools.orchestrator.Orchestrator.resolve_pr_conflicts')
     def test_resolve_conflicts_calls_orchestrator(self, mock_resolve):
         mock_resolve.return_value = {"status": "success", "message": "Resolution complete"}
         result = self.runner.invoke(cli, ['gh', 'resolve-conflicts', '--pr', '123'])
@@ -49,14 +49,14 @@ class TestModernCLI(unittest.TestCase):
         self.assertIn("Resolution complete", result.output)
         mock_resolve.assert_called_once_with(123, allow_unrelated=False, strategy=None, push=False)
 
-    @patch('tdw_services.orchestrator.Orchestrator.resolve_pr_conflicts')
+    @patch('dev_tools.orchestrator.Orchestrator.resolve_pr_conflicts')
     def test_resolve_conflicts_with_strategy_and_push(self, mock_resolve):
         mock_resolve.return_value = {"status": "success", "message": "Resolution complete"}
         result = self.runner.invoke(cli, ['gh', 'resolve-conflicts', '--pr', '123', '--strategy', 'ours', '--push'])
         self.assertEqual(result.exit_code, 0)
         mock_resolve.assert_called_once_with(123, allow_unrelated=False, strategy='ours', push=True)
 
-    @patch('tdw_services.orchestrator.Orchestrator.resolve_pr_conflicts')
+    @patch('dev_tools.orchestrator.Orchestrator.resolve_pr_conflicts')
     def test_resolve_conflicts_handles_error(self, mock_resolve):
         from utils import CLIError
         mock_resolve.side_effect = CLIError("Failed setup", code=500)
